+++ packages/box/stories/BoxStory.svelte
@@ -4,7 +4,7 @@
   export let padding = 1;
 </script>
 
-<div style="background-color: var(--pantone-red-100);">
+<div style="background-color: var(--pantone-red-100); width: max-content;">
   <Box {padding}>
     <div class="_" />
   </Box>