+++ packages/stack/stories/StackStory.svelte
@@ -1,11 +1,18 @@
 <script>
-  import Stack from "../Stack.svelte";
+  import { Stack } from "..";
 
+  export let columns = 3;
   export let space = 1;
+  export let justifyItems;
+  export let justifyContent;
+  export let alignItems;
+  export let alignContent;
 </script>
 
-<Stack {space}>
-  <div class="_" />
-  <div class="_" />
-  <div class="_" />
-</Stack>
+<div style="background-color: rgba(255,0,0,.1);">
+  <Stack {space} {justifyItems} {justifyContent} {alignItems} {alignContent}>
+    {#each Array.from({ length: columns }).fill(null) as _, i}
+      <div class="_ " />
+    {/each}
+  </Stack>
+</div>