+++ packages/tiles/Tiles.svelte
@@ -10,7 +10,7 @@
 
 <div
   class="gp-tiles"
-  style="--columns: {columns}; --space: var(--space-{space}x); --column-size: {columnSize};
+  style=" --columns: {columns}; --space: var(--space-{space}x); --column-size: {columnSize};
   --justify-items: {justifyItems}; --justify-content: {justifyContent};
   --align-items: {alignItems}; --align-content: {alignContent};">
   <slot />